requirements.txt
playwright==1.40.0
requests==2.31.0
python-dotenv==1.0.0

